App launches now, basic game runs
I fiddled to make the app runnable mby cutting out the autogenerated
code which means that we'll have to deal with some of the hiding stuff
at a later date if we have time.   A new intent starts the game
activity.   We need to make a Fragment for the description to swap out
descriptions and options menus and we need to add the rest of the
functionality to the game.

~ Ashton Spina 07/03/2017